Remove 1px space above week number in day cell
Existing CSS adds a 1px padding to the top of `.fc-content-skeleton`
in agenda and basic views. Together with the 1px top margin on
`.fc-day-grid-event`, this results in a total 2px distance between
the border and the first event.

This method was introduced in [commit b2fbdfa][1] of Aug 26, 2014 and
is based on the assumption that there are no day numbers in the
`.fc-content-skeleton` of the affected views.

However, with the new option to display week numbers in day cells
(`basicViewWeekNrPosition: 'weekNrDayCell'`), week numbers may now be
rendered in `.fc-content-skeleton`. In `basicWeek` and `basicDay`
views, the existing CSS will lead to a 1px distance between the week
number and the border above it, which may be undesirable.

This commit resolves that by reversing most of commit b2fbdfa. The
desired 2px distance above the first event is now again accomplished
by setting a 2px instead of a 1px top margin on the first
`.fc-day-grid-event` only. This in turn allows the 1px padding to the
top of `.fc-content-skeleton` to be removed.

Please review this commit carefully, since I do not know all the
reasons behind the commit it largely reverses. I am willing to discuss
alternative methods.

[1]: https://github.com/fullcalendar/fullcalendar/commit/b2fbdfa092b8bc131b61f1f3b3df98901929c66c