Simplify finding the first day of the week
As suggested by arshaw, this commit simplifies the determination of the
first day of the week (day on which the week number changes).

Note that when a custom week number function is supplied to option
`weekNumberCalculation`, the first day of the week cannot easily be
determined, so the first day of the week according to the locale
(option `lang`, possibly overridden by option `firstDay`) is taken.
The user needs to take care that the two are aligned.

When `weekNumberCalculation` is set to `'iso'`, 1 (Monday) will be
used, as per ISO 8601 definition, regardless of locale (`lang` or
`firstDay`).